backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.50.0
chromadb==0.5.0
pymupdf==1.24.0
pdfplumber==0.11.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
python-dotenv==1.0.0
pydantic==2.9.0
reportlab==4.2.0
googletrans==4.0.0-rc.1
httpx==0.27.0
